consolidate refiner head into seg_refiner.py
